I CANT TAKE IT ANYMORE I AM AT MY LIMIT

+++ src/main/java/io/github/metalcupcake5/Feesh/items/ItemType.java
@@ -15,5 +15,5 @@ public abstract class ItemType {
 
     public abstract ItemStack makeNewStack();
 
-    public abstract Map<String, Integer> getStats();
+    public abstract Map<String, Double> getStats();
 }
