+++ src/main/java/io/github/metalcupcake5/Feesh/items/Shredder.java
@@ -43,8 +43,8 @@ public class Shredder extends ItemType{
     }
 
     @Override
-    public Map<String, Integer> getStats(){
-        Map<String, Integer> stats = new HashMap<>();
+    public Map<String, Double> getStats(){
+        Map<String, Double> stats = new HashMap<>();
         return stats;
     }
 }
