+++ src/main/java/io/github/metalcupcake5/Feesh/items/xpshare.java
@@ -28,8 +28,8 @@ public class xpshare extends ItemType {
     }
 
     @Override
-    public Map<String, Integer> getStats() {
-        Map<String, Integer> stats = new HashMap<>();
+    public Map<String, Double> getStats() {
+        Map<String, Double> stats = new HashMap<>();
         return stats;
     }
 }
